+++ app/components/product/ProductCard.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import { Product } from "@/types/product";
+import { Product } from "@/app/types/product";
 
 export default function ProductCard({
   id,
